% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cssExtract.R
\name{cssExtract}
\alias{cssCharacter}
\alias{cssClass}
\alias{cssExtract}
\alias{cssId}
\alias{cssLink}
\alias{cssName}
\alias{cssNumeric}
\alias{cssSrc}
\alias{cssValue}
\title{Extract value from an html element}
\usage{
cssLink(node)

cssClass(node)

cssId(node)

cssSrc(node)

cssValue(node)

cssName(node)

cssNumeric(node, ...)

cssCharacter(node, ...)
}
\arguments{
\item{node}{An xml node}

\item{...}{Parameters passed to \code{\link{xmlValue}}}
}
\value{
All these function return a chracter string except \code{cssNumeric}
which returns a Numeric value
}
\description{
These functions have to be used in \code{\link{cssApply}} and
\code{\link{html2df}}. They aim to facilitate extraction of
 different kind of value.
}
\details{
\code{cssNumeric} extracts the numeric value of an element,
\code{cssCharacter} extracts text. \code{cssLink} extracts the url of a link.
\code{cssSrc}, \code{cssId}, \code{cssClass}, \code{cssName} and
\code{cssValue} are less usefull and extract respectively the source of an
element, its id, its css class, its name and its value (may be usefull for
input elements).
}
\examples{
doc <- "<html>
<head></head>
<body>
  <div id='character1' class='character'>
    <span class='name'>Mike</span>
    <span class='level digit'>10</span>
    <a href='http://someurl.com'>Complete profile</a>
  </div>
  <div id='character2' class='character'>
    <span class='name'>Stan</span>
    <a href='http://someurl2.com'>Complete profile</a>
  </div>
</body>
</html>"

doc <- htmlParse(doc)

# Names of the characters
cssApply(doc, ".character>.name", cssCharacter)

# Name of character1
cssApply(doc, "#character1>.name", cssCharacter)

# Urls of the profiles
cssApply(doc, ".character>a", cssLink)

# Level of characters
cssApply(doc, ".character>.level", cssNumeric)

# character 2 does not have level, we would want to have a NA value instead of nothing
cssApplyInNodeSet(doc, ".character", ".level", cssNumeric)
}

