% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cssToXpath}
\alias{cssToXpath}
\title{Translate a cssPath to an xpath.}
\usage{
cssToXpath(cssPath, prefix = "//")
}
\arguments{
\item{cssPath}{Character. A CSS path}

\item{prefix}{Character string appended at the beggining of the xpath query. Valid options
are "/", "//", "./" and ".//"}
}
\value{
A character string representing an xpath query.
}
\description{
this function translates a CSS path in an xpath query. It is used by
\code{\link{cssApply}} and you generally won't have to use it directly.
Nevertheless, it may be useful for debug if \code{\link{cssApply}} does
unexpected things.
}
\examples{
cssToXpath(".character>.name")
cssToXpath("#character1 .name")
}

