% Generated by roxygen2 (4.0.1): do not edit by hand
\name{html2df}
\alias{html2df}
\title{Convert a html document in data.frame}
\usage{
html2df(doc, obsPath, varPath, extractFun = cssCharacter)
}
\arguments{
\item{doc}{A parsed html document created with \code{\link{htmlParse}}}

\item{obsPath}{CSS path to the html elements containing the observations (see example)}

\item{varPath}{Named list containing the paths to the elements containing the values to
extract inside an observation (see example)}

\item{extractFun}{Function or vector of functions of same length than \code{varPath}. This
parameter indicates what kind of value to extract for each variable
(cssCharacter for character value, cssNumeric for numeric value, etc.)}
}
\value{
a data.frame with number of rows corresponding to the number of
elements selected by "obsPath" and number of columns equal to the length of
\code{varPath}
}
\description{
This a high level function that facilitate the creation of a data.frame from
an html document
}

