% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css.R
\name{read_css}
\alias{read_css}
\title{Parse css files into tibbles}
\usage{
read_css(file)
}
\arguments{
\item{file}{css file to parse, read with [readr::read_file()] so can be a url}
}
\value{
a tibble with columns
- rule: the name of the css rule e.g. `.some-css-class`
- setting: the name of a setting, e.g. `background-image`
- value: value of the setting, e.g. `1px solid black`
}
\description{
Parse css files into tibbles
}
