% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css.R
\name{write_css}
\alias{write_css}
\title{write css tibble into file}
\usage{
write_css(data, path, ...)
}
\arguments{
\item{data}{css tibble, e.g. generate by \code{\link[=read_css]{read_css()}}}

\item{path}{file path}

\item{...}{passed to \code{\link[readr:write_lines]{readr::write_lines()}}}
}
\description{
write css tibble into file
}
