% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css.R
\name{read_css}
\alias{read_css}
\title{read a css file into a tibble}
\usage{
read_css(file, ...)
}
\arguments{
\item{file}{css file to parse, read with \code{\link[readr:read_file]{readr::read_file()}} so can be a url}

\item{...}{\code{\%&gt;\%} \code{\link[readr:read_file]{readr::read_file()}}}
}
\value{
a tibble with columns:
\itemize{
\item rule: the name of the css rule e.g. \code{.some-css-class}
\item setting: the name of a setting, e.g. \code{background-image}
\item value: value of the setting, e.g. \code{1px solid black}
}
}
\description{
read a css file into a tibble
}
